avoid overflow in the average P (and B) frame size.  This was observed when there were  241817 P frames of ~18KB each.  Basically 4GB (32bits) isn't enough when building m2v files destined for a dual layer disk.
Fix was to use int64_t for the avg_frames array and then cast the average
size to (int).  It is highly unlikely that the average frame size will
exceed an 'int'.

+++ mjpeg_play/mplex/videostrm.hpp
@@ -86,7 +86,7 @@ public:
     unsigned int num_pictures 	;
     unsigned int num_groups 	;
     unsigned int num_frames[4] 	;
-    unsigned int avg_frames[4]  ;
+         int64_t avg_frames[4]  ;
     
     unsigned int horizontal_size;
     unsigned int vertical_size 	;
